pyproject.toml
[project]
name = "t2-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]
